Begin updating Makefile to work with git
Make live is used to make a tarball of the repository with all local
changes.

make release is used to make a release once it is tagged.

make snapshot is a work in progress; it is supposed to make a tarball of
the commit passed to it via GITREF=<commit>. The part that is missing is
I haven't decided how to name that tarball.